Modify CSV unit tests to run on travis
