+++ .travis.yml
@@ -5,8 +5,19 @@ python:
 notifications:
   slack: ooi-dev:e6WpBt9NHHw1YJjRKuYHkhYr
 
+# Setup anaconda
+before_install:
+  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+  - chmod +x miniconda.sh
+  - ./miniconda.sh -b
+  - export PATH=/home/travis/miniconda2/bin:$PATH
+  - conda update --yes conda
+
+# Install packages
 install:
-  - "pip install -r requirements.txt"
+  - conda install --yes python=$TRAVIS_PYTHON_VERSION pandas nose sqlalchemy
+  - conda install --yes -c ooi ion-functions
+  - pip install titlecase
 
 before_script:
   - "rm __init__.py"
